requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
psycopg2-binary==2.9.10
anthropic==0.40.0
python-dotenv==1.0.1
numpy>=1.26.4
pydantic==2.10.3
